% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonological_rules.R
\name{change_preGV_vowel}
\alias{change_preGV_vowel}
\title{Change /ə/ vowels preceding a glide-/ə/ sequence to appropriate quality}
\usage{
change_preGV_vowel(form)
}
\description{
Change /ə/ vowels preceding a glide-/ə/ sequence to appropriate quality
}
\details{
a. ə → ɔ / _wə
}
