% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syll_helpers.R
\name{count_sylls}
\alias{count_sylls}
\title{Count number of syllables in a candidate string}
\usage{
count_sylls(cand_str)
}
\description{
Count number of syllables in a candidate string
}
\details{
ασσσ#σ+σ → 6
}
