% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonological_rules.R
\name{delete_sfinal_schwas}
\alias{delete_sfinal_schwas}
\title{Delete stem-final /ə/ where suffixing element is vowel-initial (i.e. resolve hiatuses)}
\usage{
delete_sfinal_schwas(form)
}
\description{
Delete stem-final /ə/ where suffixing element is vowel-initial (i.e. resolve hiatuses)
}
\details{
ə → ∅ / _]V, where ] word- or morpheme-boundary
}
