% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_surface.R
\name{derive_surface}
\alias{derive_surface}
\title{Derive surface form by sequentially applying an ordered set of phonological rules}
\usage{
derive_surface(underlying_form, ordered_rules)
}
\arguments{
\item{underlying_form}{a unicode character string, e.g. 'akwelye'}

\item{ordered_rules}{character vector of phonological rules (given as names of R functions), e.g. \code{c("round_schwas", "front_initial_schwas")}}
}
\description{
Derive surface form by sequentially applying an ordered set of phonological rules
}
