% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{fconst_depc}
\alias{fconst_depc}
\title{Assign a violation to every syllable in the output with an epenthetic onset consonant}
\usage{
fconst_depc(input_str, cand_str)
}
\description{
Assign a violation to every syllable in the output with an epenthetic onset consonant
}
