% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{fconst_ident_root}
\alias{fconst_ident_root}
\title{Assign a violation if the transferred line-initial syllable was part of a line-final root form}
\usage{
fconst_ident_root(input_str, cand_str)
}
\description{
Assign a violation if the transferred line-initial syllable was part of a line-final root form
}
