% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{fconst_max_v}
\alias{fconst_max_v}
\title{#' Assign a violation to every syllable in the output that was not in the input
#'
#' @export}
\usage{
fconst_max_v(input_str, cand_str)
}
\description{
#' Assign a violation to every syllable in the output that was not in the input
#'
#' @export
}
