% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{fconst_no_transfer}
\alias{fconst_no_transfer}
\title{Assign a violation if the line-initial vowel in the input has been deleted in the output}
\usage{
fconst_no_transfer(input_str, cand_str)
}
\description{
Assign a violation if the line-initial vowel in the input has been deleted in the output
}
