% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonological_rules.R
\name{front_initial_schwas}
\alias{front_initial_schwas}
\title{Front /ə/ vowels in word-initial position}
\usage{
front_initial_schwas(form)
}
\description{
Front /ə/ vowels in word-initial position
}
\details{
ə → e / #_
}
