% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEN.R
\name{gen_all_cands}
\alias{gen_all_cands}
\title{Generate all candidates for a vector of input strings}
\usage{
gen_all_cands(syll_str, add_hinge = TRUE)
}
\description{
Generate all candidates for a vector of input strings
}
