% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txtstgs_tbl.R
\name{get_txtstgs_tbl}
\alias{get_txtstgs_tbl}
\title{Generate all rhythmically valid note combinations for n syllables}
\usage{
get_txtstgs_tbl(num_sylls)
}
\description{
Generate all rhythmically valid note combinations for n syllables
}
