% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syll_helpers.R
\name{ipa_to_cand_str}
\alias{ipa_to_cand_str}
\title{Convert transcriptions in IPA to working representations}
\usage{
ipa_to_cand_str(ipa_str)
}
\description{
Convert transcriptions in IPA to working representations
}
