% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{mconst_al_pwd_bar}
\alias{mconst_al_pwd_bar}
\title{Assign a violation to every prosodic word whose left edge is not aligned with the left edge of some bar}
\usage{
mconst_al_pwd_bar(cand_str)
}
\description{
Assign a violation to every prosodic word whose left edge is not aligned with the left edge of some bar
}
