% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{mconst_ar_lng_bar}
\alias{mconst_ar_lng_bar}
\title{Assign a violation to every syllable (whose duration is greater than a quaver) which is not immediately followed by the right edge of a bar.}
\usage{
mconst_ar_lng_bar(cand_str)
}
\description{
Assign a violation to every syllable (whose duration is greater than a quaver) which is not immediately followed by the right edge of a bar.
}
