% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{mconst_get_short}
\alias{mconst_get_short}
\title{Assign a violation where for every pair of consecutive bars, the number of syllables in the latter bar is not fewer than the number
of syllables in the former bar.}
\usage{
mconst_get_short(cand_str)
}
\description{
Assign a violation where for every pair of consecutive bars, the number of syllables in the latter bar is not fewer than the number
of syllables in the former bar.
}
