% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{mconst_no_four}
\alias{mconst_no_four}
\title{Assign a violation for every bar with 4 syllables in it}
\usage{
mconst_no_four(cand_str)
}
\description{
Assign a violation for every bar with 4 syllables in it
}
