% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{mconst_no_long}
\alias{mconst_no_long}
\title{Assign a violation where for every pair of consecutive bars, the number of syllables in the latter bar is greater than the number
of syllables in the former bar.}
\usage{
mconst_no_long(cand_str)
}
\description{
Assign a violation where for every pair of consecutive bars, the number of syllables in the latter bar is greater than the number
of syllables in the former bar.
}
