% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CON.R
\name{mconst_pwd_in_bar}
\alias{mconst_pwd_in_bar}
\title{Assign a violation to every bar which does not contain exactly one prosodic word}
\usage{
mconst_pwd_in_bar(cand_str)
}
\description{
Assign a violation to every bar which does not contain exactly one prosodic word
}
