% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduplicate.R
\name{reduplicate}
\alias{reduplicate}
\title{Determine reduplicant from base and suffix accordingly}
\usage{
reduplicate(base, type = "2-3")
}
\arguments{
\item{base}{a unicode character string}

\item{type}{type of reduplication to perform (e.g. 2-3)}
}
\description{
Determine reduplicant from base and suffix accordingly
}
