% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonological_rules.R
\name{round_schwas}
\alias{round_schwas}
\title{Round /ə/ vowels preceding a rounded consonant /Cʷ/ or labial approximant /w/}
\usage{
round_schwas(form)
}
\description{
Round /ə/ vowels preceding a rounded consonant /Cʷ/ or labial approximant /w/
}
\details{
a. ə → u / Cʷ_

b. ə → u / w_
}
